New component Article and some fix
