+++ src/components/Article/Article.jsx
@@ -1,44 +1,31 @@
 import React from 'react';
-import Typography from '@material-ui/core/Typography';
-import Card from '@material-ui/core/Card';
-import Button from '@material-ui/core/Button';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
+import { connect } from "react-redux";
 import { useStyles } from './StyleArticle';
+import Button from '@material-ui/core/Button';
+import Banner from '../Banner';
 
 const Article = (props) => {
-
-  const { id, brand, name, price, vendor, condition, description, image } = props;
-
-  const classes = useStyles();
-
+  const { id } = props.match.params;
+  //onsole.log(id);
+  console.log('props: ',props);
+  console.log('props.match: ',props.match);
+  console.log('props.match.params: ', props.match.params);
   return (
-    <Card className={classes.card}>
-      <CardMedia
-        className={classes.cardMedia}
-        image= {image}
-        title="Image title"
-      />
-      <CardContent className={classes.cardContent}>
-        <Typography
-          gutterBottom
-          variant="h5"
-          component="h2">
-          {name}
-        </Typography>
-        <Typography> {description} </Typography>
-      </CardContent>
-      <CardActions>
-        <Button size="small" color="primary">
-          Favoritos
-         </Button>
-        <Button size="small" color="primary">
-          Agregar a carrito
-        </Button>
-      </CardActions>
-    </Card>
-  );
-
+    <main>
+      <Banner />
+      <h1> {id} </h1>
+      <Button onClick={() => props.history.goBack()} color="primary">
+        Volver
+      </Button>
+    </main>
+  )
 }
-export default Article;
+
+
+const mapStateToProps = state => {
+  return {
+    articles: state.articles,
+  };
+};
+
+export default connect(mapStateToProps, null)(Article);
