+++ src/components/CardArticle/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CardArticle.jsx";
